# packages/client/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Bandroom.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5180/"
    }
  ],
  "paths": {
...
    "/bands": {
      "post": {
        "tags": [
          "bands"
        ],
        "summary": "Create a band (creator becomes admin)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BandDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "bands"
        ],
        "summary": "The caller's bands",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BandSummaryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bands/{bandId}": {
      "get": {
        "tags": [
          "bands"
        ],
        "summary": "Band detail + members",
        "parameters": [
          {
            "name": "bandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BandDetailResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "bands"
        ],
        "summary": "Update band settings (admin)",
        "parameters": [
          {
            "name": "bandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/bands/{bandId}/channels": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Channels (#general always exists)",
        "parameters": [
          {
            "name": "bandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create a channel",
        "parameters": [
          {
            "name": "bandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/bands/{bandId}/channels/{channelId}/messages": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Messages, newest page first",
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "bandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesPage"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send a message",
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "bandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BandDetailResponse": {
        "required": [
          "band",
          "members"
        ],
        "type": "object",
        "properties": {
          "band": {
            "$ref": "#/components/schemas/BandResponse"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            }
          }
        }
      },
      "BandPlan": {
        "enum": [
          "free",
          "pro"
        ]
      },
      "BandResponse": {
        "required": [
          "id",
          "name",
          "timeZone",
          "quorum",
          "defaultPracticeSlot",
          "rehearsalSpace",
          "plan",
          "storageUsedBytes",
          "storageQuotaBytes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "quorum": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "defaultPracticeSlot": {
            "$ref": "#/components/schemas/PracticeSlot"
          },
          "rehearsalSpace": {
            "type": [
              "null",
              "string"
            ]
          },
          "plan": {
            "$ref": "#/components/schemas/BandPlan"
          },
          "storageUsedBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int64"
          },
          "storageQuotaBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BandRole": {
        "enum": [
          "member",
          "admin"
        ]
      },
      "BandSummaryResponse": {
        "required": [
          "id",
          "name",
          "role"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/BandRole"
          }
        }
      },
      "ChannelResponse": {
        "required": [
          "id",
          "name",
          "isDefault",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Instant"
          }
        }
      },
...
      "CreateBandRequest": {
        "required": [
          "name",
          "timeZone"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "timeZone": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateChannelRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Instant": {
        "type": "string",
        "format": "date-time"
      },
...
      "MemberResponse": {
        "required": [
          "membershipId",
          "userId",
          "displayName",
          "role",
          "instrument",
          "joinedAt"
        ],
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/BandRole"
          },
          "instrument": {
            "type": [
              "null",
              "string"
            ]
          },
          "joinedAt": {
            "$ref": "#/components/schemas/Instant"
          }
        }
      },
...
      "MessageResponse": {
        "required": [
          "id",
          "authorMembershipId",
          "authorName",
          "body",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "authorMembershipId": {
            "type": "string",
            "format": "uuid"
          },
          "authorName": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Instant"
          }
        }
      },
      "MessagesPage": {
        "required": [
          "items",
          "nextBefore"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            }
          },
          "nextBefore": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PracticeSlot": {
        "enum": [
          "afternoon",
          "evening"
        ]
      },
...
      "SendMessageRequest": {
        "required": [
          "body"
        ],
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        }
      },
...
      "UpdateBandRequest": {
        "required": [
          "name",
          "timeZone",
          "quorum",
          "defaultPracticeSlot",
          "rehearsalSpace"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "timeZone": {
            "type": [
              "null",
              "string"
            ]
          },
          "quorum": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "defaultPracticeSlot": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PracticeSlot"
              }
            ]
          },
          "rehearsalSpace": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}